+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file. **I Tried.*
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). **I Tried.**
 
+## [0.2.2-unreleased] 2019-01-30
+
+Skipped Versions to fix the wrong version numbers.
+
+### Added
+
+- New Custom Reports:
+  - Steam
+  - Apple
+- Proper Logging
+
+### Fixed
+
+- Old Custom Report Commands from all capitalized to capitalized first letter
+- "Google Sheets "UNAUTHENTICATED" when left idle for some time" when creating a new entry
+- Google Sheets running out of columns
+
 ## [0.1.6.2-unreleased] 2018-12-28
 
 HotFix